Move the client package contents to the root package.
