+++ firebase-web/src/main/java/io/spine/web/firebase/query/QueryRecord.java
@@ -25,9 +25,9 @@ import io.spine.client.Query;
 import io.spine.client.QueryResponse;
 import io.spine.json.Json;
 import io.spine.protobuf.AnyPacker;
-import io.spine.web.firebase.client.FirebaseClient;
+import io.spine.web.firebase.FirebaseClient;
+import io.spine.web.firebase.NodeValue;
 import io.spine.web.firebase.client.NodePath;
-import io.spine.web.firebase.client.NodeValue;
 
 import java.util.List;
 import java.util.concurrent.CompletionStage;
