+++ firebase-web/src/test/java/io/spine/web/firebase/NodeValueTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.web.firebase.client;
+package io.spine.web.firebase;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
