+++ firebase-web/src/test/java/io/spine/web/firebase/given/FirebaseSubscriptionBridgeTestEnv.java
@@ -35,7 +35,7 @@ import io.spine.core.ResponseVBuilder;
 import io.spine.core.UserId;
 import io.spine.core.UserIdVBuilder;
 import io.spine.testing.client.TestActorRequestFactory;
-import io.spine.web.firebase.client.FirebaseClient;
+import io.spine.web.firebase.FirebaseClient;
 import io.spine.web.firebase.subscription.FirebaseSubscriptionBridge;
 
 import javax.servlet.ServletResponse;
