+++ firebase-web/src/test/java/io/spine/web/firebase/query/FirebaseQueryBridgeTest.java
@@ -28,7 +28,7 @@ import io.spine.client.Query;
 import io.spine.client.QueryFactory;
 import io.spine.client.grpc.QueryServiceGrpc.QueryServiceImplBase;
 import io.spine.testing.client.TestActorRequestFactory;
-import io.spine.web.firebase.client.FirebaseClient;
+import io.spine.web.firebase.FirebaseClient;
 import io.spine.web.firebase.given.FirebaseQueryMediatorTestEnv.TestQueryService;
 import io.spine.web.firebase.subscription.given.HasChildren;
 import io.spine.web.query.QueryProcessingResult;
