+++ firebase-web/src/test/java/io/spine/web/firebase/subscription/FirebaseSubscribeResultTest.java
@@ -23,8 +23,8 @@ package io.spine.web.firebase.subscription;
 import io.spine.client.Subscription;
 import io.spine.client.Topic;
 import io.spine.client.TopicFactory;
+import io.spine.web.firebase.NodePaths;
 import io.spine.web.firebase.client.NodePath;
-import io.spine.web.firebase.client.NodePaths;
 import io.spine.web.subscription.result.SubscribeResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
