+++ firebase-web/src/test/java/io/spine/web/firebase/subscription/FirebaseSubscriptionBridgeTest.java
@@ -25,7 +25,7 @@ import io.spine.client.Topic;
 import io.spine.client.TopicFactory;
 import io.spine.client.grpc.QueryServiceGrpc.QueryServiceImplBase;
 import io.spine.core.Response;
-import io.spine.web.firebase.client.FirebaseClient;
+import io.spine.web.firebase.FirebaseClient;
 import io.spine.web.subscription.result.SubscribeResult;
 import io.spine.web.subscription.result.SubscriptionCancelResult;
 import io.spine.web.subscription.result.SubscriptionKeepUpResult;
