+++ firebase-web/src/test/java/io/spine/web/firebase/subscription/QueryRecordTest.java
@@ -21,10 +21,10 @@
 package io.spine.web.firebase.subscription;
 
 import io.spine.client.QueryResponse;
-import io.spine.web.firebase.client.FirebaseClient;
+import io.spine.web.firebase.FirebaseClient;
+import io.spine.web.firebase.NodePaths;
+import io.spine.web.firebase.NodeValue;
 import io.spine.web.firebase.client.NodePath;
-import io.spine.web.firebase.client.NodePaths;
-import io.spine.web.firebase.client.NodeValue;
 import io.spine.web.firebase.given.Book;
 import io.spine.web.firebase.subscription.given.HasChildren;
 import org.junit.jupiter.api.DisplayName;
