+++ integration-tests/web-tests/src/main/java/io/spine/web/test/given/Application.java
@@ -23,15 +23,15 @@ package io.spine.web.test.given;
 import io.spine.server.BoundedContext;
 import io.spine.server.CommandService;
 import io.spine.server.QueryService;
+import io.spine.web.firebase.DatabaseUrls;
+import io.spine.web.firebase.FirebaseClient;
+import io.spine.web.firebase.FirebaseCredentials;
 import io.spine.web.firebase.client.DatabaseUrl;
-import io.spine.web.firebase.client.DatabaseUrls;
-import io.spine.web.firebase.client.FirebaseClient;
-import io.spine.web.firebase.client.FirebaseCredentials;
 
 import java.io.InputStream;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static io.spine.web.firebase.client.FirebaseClientFactory.restClient;
+import static io.spine.web.firebase.FirebaseClientFactory.restClient;
 
 /**
  * A test Spine application.
